requirements.txt
fastapi
uvicorn
chromadb
sentence-transformers
pypdf2
python-docx
pytesseract
pillow
transformers
torch